+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.2.0
+- Added new color palettes: Spectrum, Rainbow Spectrum, Nightfall, Aquaflow, Luminarium, Monochrome, and Solarized.
+- Enhanced user experience by improving the language of the opacity slider. Renamed "Less Opacity"/"More Opacity" to "Transparent"/"Opaque".
+
 ## 2.1.0
 - Add option to make edges round
 - Fix language detection for "Never highlight indent as error" option